Update start and stop functions in .c file
This also fixes the start and stop condtions. There are still timing
issues with start and stop conditions which will be fixed (hopefully) in
the next commit. This issue is that the SDA and SCL transitions in the
start and stop conditions are probably too close together, and it is
difficult to create delays without accidentally shifting the transmitted
bits.